converted sigma to a float value

+++ RZEffects/Effects/RZBlurEffect.h
@@ -9,7 +9,7 @@
 
 @interface RZBlurEffect : RZEffect
 
-@property (assign, nonatomic) GLint sigma;
+@property (assign, nonatomic) GLfloat sigma;
 
 + (instancetype)effect;
